% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamictable.R
\name{colorizeString}
\alias{colorizeString}
\title{Colorize String}
\usage{
colorizeString(x, min = 128, max = 240, r = NA, g = NA, b = NA,
  as.html = TRUE)
}
\arguments{
\item{x}{The string(s) to be hashed}

\item{min}{Default 128. The minimum value (0-256) any RGB value can
have}

\item{max}{Default 240. Maximum value an RGB value can have}

\item{r}{Explicit value to set for the red channel. Ignores "min"
and "max". Combined with "g" and "b" allows some control over
the hues generated during hashing.}

\item{g}{Explict green channel value}

\item{b}{Explicit blue channel value}

\item{as.html}{When TRUE (default) the returned value will be
hexadecimal strings suitable (eg "#aa318f"). If false, a three
column ("R","G","B") array will be returned, one row per input
string.}
}
\description{
Takes one or more strings as input, and generates a 24-bit
hexadecimal color based on hashing the string
}
\examples{

colorizeString(c("quick fox","QUICK fox", "lazy dog"))
colorizeString(c("quick fox","QUICK fox", "lazy dog"), as.html=FALSE)

}

