% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamictable.R
\docType{class}
\name{DynamicTable-class}
\alias{DynamicTable}
\alias{DynamicTable-class}
\title{Dynamic Table Object}
\description{
This is the underlying ReferenceClass object supporting
\code{dynamictable}. While it could be used directly, it is
presumed that in most cases users will be calling the function
\code{dynamictable()}

See \code{?dynamic} for a full parameter listing
}
\section{Fields}{

\describe{
\item{\code{data}}{The primary data, held as a data.frame (either as
provided, or converted to one)}

\item{\code{cols}}{Character vector of columns to display}

\item{\code{colInfo}}{List holding supporting structures for building
columns in the HTML table}

\item{\code{autoCols}}{Logical value, TRUE if columns were simple
extracted from the supplied object, FALSE if user has specified
particular columns to use.}

\item{\code{userStyle}}{Character vector holding any user-supplied CSS
styles}

\item{\code{style}}{Character vector, programatically-generated styles}

\item{\code{file}}{File path for output}

\item{\code{options}}{A list holding both user-supplied and generated
'primary' settings. See \code{?dynamic} for full details on
options}

\item{\code{params}}{A list, similar to \code{options} - See
\code{?dynamic} for full details}

\item{\code{flags}}{A list of assorted backend configuration flags}

\item{\code{title}}{The title (upper left browser window frame)}

\item{\code{header}}{Optional text to put above the table}

\item{\code{footer}}{Optional text for below the table}

\item{\code{sortby}}{Optional column for an initial sort}
}}

