% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamictable.R
\name{is.def}
\alias{is.def}
\title{Is Defined}
\usage{
is.def(x)
}
\arguments{
\item{x}{The object being tested}
}
\value{
TRUE or FALSE
}
\description{
Convenience method, returns FALSE if the object is null, an
uninitialized RefClass field, a zero-length matrix or all NAs.
}

