% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamictable.R
\name{parenRE}
\alias{parenRE}
\title{Parenthetical Regular Expression}
\usage{
parenRE(RegExp, text)
}
\arguments{
\item{RegExp}{The regular expression, presumably with one or more
parenthetical components}

\item{text}{The string to match. NOT vectorized}
}
\value{
If no matches found \code{NA}; Otherwise a character vector
    with one element per parenthetical capture
}
\description{
Simplifies parenthetical extraction of matches from regular expressions
}
\examples{

parenRE("(quick|lazy)\\\\b.+(fox|DOG)", "Quick brown fox")
## Word boundary (\\b) prevents matching to "quickest":
parenRE("(quick|lazy)\\\\b.+(fox|DOG)", "Quickest brown fox")

}

