% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texteditorplatforms}
\alias{texteditorplatforms}
\title{Compatibility grid of text editor software with major operating systems}
\format{A data.frame with 74 rows and 6 variables representing
    operating systems: "Windows", "OS.X", "Linux", "BSD", "Unix"
    and "OpenVMS"}
\usage{
texteditorplatforms
}
\description{
Extracted from Wikipedia article
  "Comparison of text editors". Intended to showcase factor
  filtering in dynamictable. The data are very regular with three
  levels ("Yes", "No", "Partial") plus \code{NA}.
}
\examples{

dynamictable(texteditorplatforms)
}
\references{
\url{https://en.wikipedia.org/wiki/Comparison_of_text_editors#Cross-platform}
}
\keyword{datasets}

