% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typesystemcomparison}
\alias{typesystemcomparison}
\title{Table highlighting the type systems used by various programming languages}
\format{A data.frame with 41 rows and 4 variables
\itemize{
  \item static.dynamic: Are safety checks performed at compilation (static) or runtime (dynamic)?
  \item strong.weak: Is it easy to violate type constraints (weak) or will errors be generated (strong). Terminology is contentious here.
  \item safety: Another contentious category.
  \item nominative.structural: Is the system, nominative, structural, duck or some combination?
}}
\usage{
typesystemcomparison
}
\description{
Extracted from Wikipedia article
  "Comparison of type systems". Intended to showcase factor
  filtering in dynamictable. The data consist of relatively clean
  factor levels.
}
\examples{

dynamictable(typesystemcomparison)
}
\references{
\url{https://en.wikipedia.org/wiki/Comparison_of_type_systems}
}
\keyword{datasets}

